+++ bootware.ps1
@@ -878,7 +878,7 @@ Function UpdateCompletion($Version) {
 
 # Print Bootware version string.
 Function Version() {
-    Write-Output 'Bootware 0.5.2'
+    Write-Output 'Bootware 0.5.3'
 }
 
 # Convert path to WSL relative path.